requirements.txt
Django==4.2.5
djangorestframework==3.14.0
jsonschema==4.19.1
PyYAML==6.0.1
